[ticket/15747] Change parameter depth to boolean

+++ phpBB/phpbb/storage/adapter/local.php
@@ -73,10 +73,17 @@ class local implements adapter_interface, stream_interface
 	 * This is for those who have problems storing a large number of files in
 	 * a single directory.
 	 * More info: https://tracker.phpbb.com/browse/PHPBB3-15371
-	 *
+	 */
+
+	/*
+	 * @var bool subfolders
+	 */
+	protected $subfolders;
+
+	/*
 	 * @var int dir_depth
 	 */
-	protected $dir_depth;
+	protected $dir_depth = 2;
 
 	/**
 	 * Constructor
@@ -101,7 +108,7 @@ class local implements adapter_interface, stream_interface
 
 		$this->path = $options['path'];
 		$this->root_path = $this->phpbb_root_path . $options['path'];
-		$this->dir_depth = (int) $options['depth'];
+		$this->subfolders = (bool) $options['subfolders'];
 	}
 
 	/**
@@ -249,17 +256,20 @@ class local implements adapter_interface, stream_interface
 	 */
 	protected function remove_empty_dirs($path)
 	{
-		$dirpath = dirname($this->root_path . $path);
-		$filepath = dirname($this->root_path . $this->get_path($path) . $this->get_filename($path));
-		$path = filesystem_helper::make_path_relative($filepath, $dirpath);
-
-		do
+		if ($this->subfolders)
 		{
-			$parts = explode('/', $path);
-			$parts = array_slice($parts, 0, -1);
-			$path = implode('/', $parts);
+			$dirpath = dirname($this->root_path . $path);
+			$filepath = dirname($this->root_path . $this->get_path($path) . $this->get_filename($path));
+			$path = filesystem_helper::make_path_relative($filepath, $dirpath);
+
+			do
+			{
+				$parts = explode('/', $path);
+				$parts = array_slice($parts, 0, -1);
+				$path = implode('/', $parts);
+			}
+			while ($path && @rmdir($dirpath . '/' . $path));
 		}
-		while ($path && @rmdir($dirpath . '/' . $path));
 	}
 
 	/**
@@ -271,21 +281,22 @@ class local implements adapter_interface, stream_interface
 	protected function get_path($path)
 	{
 		$dirname = dirname($path);
+		$dirname = ($dirname != '.') ? $dirname . DIRECTORY_SEPARATOR : '';
 
-		$hash = md5(basename($path));
-
-		$parts = str_split($hash, 2);
-		$parts = array_slice($parts, 0, $this->dir_depth);
-
-		// Create path
-		$path = $dirname . DIRECTORY_SEPARATOR;
-
-		if (!empty($parts))
+		if ($this->subfolders)
 		{
-			$path .= implode(DIRECTORY_SEPARATOR, $parts) . DIRECTORY_SEPARATOR;
+			$hash = md5(basename($path));
+
+			$parts = str_split($hash, 2);
+			$parts = array_slice($parts, 0, $this->dir_depth);
+
+			if (!empty($parts))
+			{
+				$dirname .= implode(DIRECTORY_SEPARATOR, $parts) . DIRECTORY_SEPARATOR;
+			}
 		}
 
-		return $path;
+		return $dirname;
 	}
 
 	/**
